# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface UploadResponse {
  jobId: string;
  message: string;
}

export type AnalysisStatusStatus =
  (typeof AnalysisStatusStatus)[keyof typeof AnalysisStatusStatus];

export const AnalysisStatusStatus = {
  queued: "queued",
  uploading: "uploading",
  extracting_audio: "extracting_audio",
  extracting_frames: "extracting_frames",
  transcribing: "transcribing",
  analyzing_visual: "analyzing_visual",
  analyzing_audio: "analyzing_audio",
  analyzing_content: "analyzing_content",
  generating_seo: "generating_seo",
  generating_subtitles: "generating_subtitles",
  complete: "complete",
  error: "error",
} as const;

export interface AnalysisStatus {
  jobId: string;
  status: AnalysisStatusStatus;
  progress: number;
  currentStep: string;
  error?: string;
}

export interface MetricItem {
  level: string;
  numeric?: number;
  assessment: string;
  suggestions?: string[];
  effect?: string;
}

export interface QualityResult {
  score: number;
  lighting?: MetricItem;
  brightness?: MetricItem;
  contrast?: MetricItem;
  sharpness?: MetricItem;
  stability?: MetricItem;
  colorBalance?: MetricItem;
  background?: MetricItem;
  framing?: MetricItem;
  audioVolume?: MetricItem;
  audioClarity?: MetricItem;
  backgroundNoise?: MetricItem;
  fillerWords?: MetricItem;
  pacing?: MetricItem;
  suggestions: string[];
}

export type WeakSectionColor =
  (typeof WeakSectionColor)[keyof typeof WeakSectionColor];

export const WeakSectionColor = {
  red: "red",
  yellow: "yellow",
} as const;

export interface WeakSection {
  start: string;
  end: string;
  text: string;
  color: WeakSectionColor;
  reason: string;
}

export interface ContentResult {
  hookScore: number;
  hooks: string[];
  weakSections: WeakSection[];
  improvements: string[];
  audience: string;
  problemSolved: string;
}

export interface HashtagItem {
  tag: string;
  effect: string;
}

export interface TimestampItem {
  time: string;
  label: string;
}

export interface SeoResult {
  titles: string[];
  description: string;
  hashtags: HashtagItem[];
  timestamps: TimestampItem[];
}

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
}

export interface SubtitlesResult {
  transcript: TranscriptSegment[];
  fullText: string;
  translatedTranscript?: TranscriptSegment[];
  translatedLanguage?: string;
}

export interface AnalysisResult {
  jobId: string;
  platform: string;
  quality: QualityResult;
  content: ContentResult;
  seo: SeoResult;
  subtitles: SubtitlesResult;
}

export type ExportRequestResolution =
  (typeof ExportRequestResolution)[keyof typeof ExportRequestResolution];

export const ExportRequestResolution = {
  "240p": "240p",
  "480p": "480p",
  "720p": "720p",
  "1080p": "1080p",
  "4k": "4k",
} as const;

export interface ExportRequest {
  resolution: ExportRequestResolution;
  removeFillerWords?: boolean;
  includeSubtitles?: boolean;
}

export interface ExportResponse {
  downloadUrl: string;
  filename: string;
}

export type UploadVideoBodyPlatform =
  (typeof UploadVideoBodyPlatform)[keyof typeof UploadVideoBodyPlatform];

export const UploadVideoBodyPlatform = {
  youtube_long: "youtube_long",
  youtube_shorts: "youtube_shorts",
  tiktok: "tiktok",
  instagram: "instagram",
  linkedin: "linkedin",
  x: "x",
} as const;

export type UploadVideoBody = {
  platform: UploadVideoBodyPlatform;
  translateSubtitles?: boolean;
  subtitleLanguage?: string;
  replaceAudio?: boolean;
  audioLanguage?: string;
};