# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentResult } from "./contentResult";
import type { QualityResult } from "./qualityResult";
import type { SeoResult } from "./seoResult";
import type { SubtitlesResult } from "./subtitlesResult";

export interface AnalysisResult {
  jobId: string;
  platform: string;
  quality: QualityResult;
  content: ContentResult;
  seo: SeoResult;
  subtitles: SubtitlesResult;
}