# lib/api-zod/src/generated/types/analysisStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatusStatus } from "./analysisStatusStatus";

export interface AnalysisStatus {
  jobId: string;
  status: AnalysisStatusStatus;
  progress: number;
  currentStep: string;
  error?: string;
}