# lib/api-zod/src/generated/types/exportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportRequestResolution } from "./exportRequestResolution";

export interface ExportRequest {
  resolution: ExportRequestResolution;
  removeFillerWords?: boolean;
  includeSubtitles?: boolean;
}