# lib/api-zod/src/generated/types/exportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ExportResponse {
  downloadUrl: string;
  filename: string;
}