# lib/api-zod/src/generated/types/metricItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricItem {
  level: string;
  numeric?: number;
  assessment: string;
  suggestions?: string[];
  effect?: string;
}