# lib/api-zod/src/generated/types/qualityResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { MetricItem } from "./metricItem";

export interface QualityResult {
  score: number;
  lighting?: MetricItem;
  brightness?: MetricItem;
  contrast?: MetricItem;
  sharpness?: MetricItem;
  stability?: MetricItem;
  colorBalance?: MetricItem;
  background?: MetricItem;
  framing?: MetricItem;
  audioVolume?: MetricItem;
  audioClarity?: MetricItem;
  backgroundNoise?: MetricItem;
  fillerWords?: MetricItem;
  pacing?: MetricItem;
  suggestions: string[];
}