# lib/api-zod/src/generated/types/subtitlesResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { TranscriptSegment } from "./transcriptSegment";

export interface SubtitlesResult {
  transcript: TranscriptSegment[];
  fullText: string;
  translatedTranscript?: TranscriptSegment[];
  translatedLanguage?: string;
}