# lib/api-zod/src/generated/types/weakSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayTabs Video Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { WeakSectionColor } from "./weakSectionColor";

export interface WeakSection {
  start: string;
  end: string;
  text: string;
  color: WeakSectionColor;
  reason: string;
}